app/src/main/java/com/ser210/cyr/clevernotes2/EntityClasses/Tag.java
/**
 * Authors: Jake Cyr and Yishuo Tang
 * Date: 27 April 2016
 * Description: This is a tag entity class that stores and returns data for a single tag.
 */

package com.ser210.cyr.clevernotes2.EntityClasses;

public class Tag {

    private String tag, created;
    private long id;

    public String getTag() {
        return tag;
    }

    public void setTag(String tag) {
        this.tag = tag;
    }

    public String getCreated() {
        return created;
    }

    public void setCreated(String created) {
        this.created = created;
    }

    public long getId() {
        return id;
    }

    public void setId(long id) {
        this.id = id;
    }
}
